Refactored GLB utils
Moved the GLB methods into a single module.
